// src/models/api/interrogationMetadata.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * queen-api
 * API for Queen/Stromae
 * OpenAPI spec version: 5.0.0-rc
 */
import type { InterrogationMetadataContext } from './interrogationMetadataContext'
import type { Logos } from './logos'
import type { MetadataVariable } from './metadataVariable'
import type { SchemaPersonalization } from './schema.personalization/schemaPersonalization'

export interface InterrogationMetadata {
  context: InterrogationMetadataContext
  personalization?: SchemaPersonalization
  label: string
  objectives: string
  variables?: MetadataVariable[]
  logos?: Logos
}